ARROW-14243: [C++] Split vector_sort.cc
Move array sorting to a separate translation unit.
This should make vector_sort.cc slightly less heavy to compile.
