+++ cpp/src/arrow/CMakeLists.txt
@@ -408,6 +408,7 @@ if(ARROW_COMPUTE)
        compute/kernels/scalar_validity.cc
        compute/kernels/scalar_if_else.cc
        compute/kernels/util_internal.cc
+       compute/kernels/vector_array_sort.cc
        compute/kernels/vector_hash.cc
        compute/kernels/vector_nested.cc
        compute/kernels/vector_replace.cc
